requirements.txt
fastapi[all]
uvicorn
sqlalchemy
httpx
alembic
psycopg2-binary
python-dotenv